LayoutTests/fast/js/dfg-div-by-neg1-and-then-or-zero-interesting-reg-alloc-expected.txt

Tests that the (x/-1)|0 case in the DFG is handled correctly even when there is some interesting register allocation going on.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(0, 0, 0, -1) is 0
PASS foo(0, 0, 1, -1) is -1
PASS foo(0, 0, 2, -1) is -2
PASS foo(0, 0, 3, -1) is -3
PASS foo(0, 0, 4, -1) is -4
PASS foo(0, 0, 5, -1) is -5
PASS foo(0, 0, 6, -1) is -6
PASS foo(0, 0, 7, -1) is -7
PASS foo(0, 0, 8, -1) is -8
PASS foo(0, 0, 9, -1) is -9
PASS foo(0, 0, 10, -1) is -10
PASS foo(0, 0, 11, -1) is -11
PASS foo(0, 0, 12, -1) is -12
PASS foo(0, 0, 13, -1) is -13
PASS foo(0, 0, 14, -1) is -14
PASS foo(0, 0, 15, -1) is -15
PASS foo(0, 0, 16, -1) is -16
PASS foo(0, 0, 17, -1) is -17
PASS foo(0, 0, 18, -1) is -18
PASS foo(0, 0, 19, -1) is -19
PASS foo(0, 0, 20, -1) is -20
PASS foo(0, 0, 21, -1) is -21
PASS foo(0, 0, 22, -1) is -22
PASS foo(0, 0, 23, -1) is -23
PASS foo(0, 0, 24, -1) is -24
PASS foo(0, 0, 25, -1) is -25
PASS foo(0, 0, 26, -1) is -26
PASS foo(0, 0, 27, -1) is -27
PASS foo(0, 0, 28, -1) is -28
PASS foo(0, 0, 29, -1) is -29
PASS foo(0, 0, 30, -1) is -30
PASS foo(0, 0, 31, -1) is -31
PASS foo(0, 0, 32, -1) is -32
PASS foo(0, 0, 33, -1) is -33
PASS foo(0, 0, 34, -1) is -34
PASS foo(0, 0, 35, -1) is -35
PASS foo(0, 0, 36, -1) is -36
PASS foo(0, 0, 37, -1) is -37
PASS foo(0, 0, 38, -1) is -38
PASS foo(0, 0, 39, -1) is -39
PASS foo(0, 0, 40, -1) is -40
PASS foo(0, 0, 41, -1) is -41
PASS foo(0, 0, 42, -1) is -42
PASS foo(0, 0, 43, -1) is -43
PASS foo(0, 0, 44, -1) is -44
PASS foo(0, 0, 45, -1) is -45
PASS foo(0, 0, 46, -1) is -46
PASS foo(0, 0, 47, -1) is -47
PASS foo(0, 0, 48, -1) is -48
PASS foo(0, 0, 49, -1) is -49
PASS foo(0, 0, 50, -1) is -50
PASS foo(0, 0, 51, -1) is -51
PASS foo(0, 0, 52, -1) is -52
PASS foo(0, 0, 53, -1) is -53
PASS foo(0, 0, 54, -1) is -54
PASS foo(0, 0, 55, -1) is -55
PASS foo(0, 0, 56, -1) is -56
PASS foo(0, 0, 57, -1) is -57
PASS foo(0, 0, 58, -1) is -58
PASS foo(0, 0, 59, -1) is -59
PASS foo(0, 0, 60, -1) is -60
PASS foo(0, 0, 61, -1) is -61
PASS foo(0, 0, 62, -1) is -62
PASS foo(0, 0, 63, -1) is -63
PASS foo(0, 0, 64, -1) is -64
PASS foo(0, 0, 65, -1) is -65
PASS foo(0, 0, 66, -1) is -66
PASS foo(0, 0, 67, -1) is -67
PASS foo(0, 0, 68, -1) is -68
PASS foo(0, 0, 69, -1) is -69
PASS foo(0, 0, 70, -1) is -70
PASS foo(0, 0, 71, -1) is -71
PASS foo(0, 0, 72, -1) is -72
PASS foo(0, 0, 73, -1) is -73
PASS foo(0, 0, 74, -1) is -74
PASS foo(0, 0, 75, -1) is -75
PASS foo(0, 0, 76, -1) is -76
PASS foo(0, 0, 77, -1) is -77
PASS foo(0, 0, 78, -1) is -78
PASS foo(0, 0, 79, -1) is -79
PASS foo(0, 0, 80, -1) is -80
PASS foo(0, 0, 81, -1) is -81
PASS foo(0, 0, 82, -1) is -82
PASS foo(0, 0, 83, -1) is -83
PASS foo(0, 0, 84, -1) is -84
PASS foo(0, 0, 85, -1) is -85
PASS foo(0, 0, 86, -1) is -86
PASS foo(0, 0, 87, -1) is -87
PASS foo(0, 0, 88, -1) is -88
PASS foo(0, 0, 89, -1) is -89
PASS foo(0, 0, 90, -1) is -90
PASS foo(0, 0, 91, -1) is -91
PASS foo(0, 0, 92, -1) is -92
PASS foo(0, 0, 93, -1) is -93
PASS foo(0, 0, 94, -1) is -94
PASS foo(0, 0, 95, -1) is -95
PASS foo(0, 0, 96, -1) is -96
PASS foo(0, 0, 97, -1) is -97
PASS foo(0, 0, 98, -1) is -98
PASS foo(0, 0, 99, -1) is -99
PASS successfullyParsed is true

TEST COMPLETE
